CamMath.jl

A package for image analysis of backscattered light

Table of contents

CamMath.jl

Table of contents

Bernoulli number

Bernoulli number

CamMath.bernoulliB — Method

bernoulliB(n::T [; msg=true]) where {T<:Integer}

Bernoulli numbers of index n are defined by the recurrence relation

\[ B_n = - \frac{1}{n+1}\sum_{k=0}^{n-1}\frac{(n+1)!}{k!(n+1-k)}B_k,\]

with $B_0=1$ and $B_1=-1/2$. Starting at $B_0$ is called the even index convention (B_{2n+1}=0\ \rm{for}\ n>1)``.

Integer-overflow protection (IOP): for n > 35 the output is autoconverted to Rational{BigInt}. By default the capture message is activated: "Warning: IOP - bernoulliB converted to Rational{BigInt}".

Examples:

julia> o = [bernoulliB(n) for n=0:5]; println(o)
Rational{Int64}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1]

julia> bernoulliB(60)
Warning: IOP - bernoulliB converted to Rational{BigInt}
-1215233140483755572040304994079820246041491//56786730

julia> n = 60;
julia> bernoulliB(n) == bernoulliB_array(n)[end]             
true

CamMath.bernoulliB_array — Method

bernoulliB_array(nmax::T [; msg=true]) where {T<:Integer}

Bernoulli number array $[B_0,\cdots\ B_nmax]$. NB.: nmax is the index of the highest Bernoulli number of the array (not the array length).

Examples:

julia> o = bernoulliB_array(8); println(o)
Rational{Int64}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1, 1//42, 0//1, -1//30]

julia> n = 60; msg = false;
julia>  bernoulliB(n; msg) == bernoulliB_array(n; msg)[end]            
true