CamMath.jl

Mathematics library with integer-overload protection

Table of contents

CamMath.jl

Table of contents

Bernoulli number

Factorial

Faulhaber polynomial

HarmonicNumber

Pascal triangle

Bernoulli number

CamMath.bernoulliB — Method

bernoulliB(n::T [; msg=true]) where {T<:Integer}

Bernoulli numbers of index n are defined by the recurrence relation

\[ B_n = - \frac{1}{n+1}\sum_{k=0}^{n-1}\frac{(n+1)!}{k!(n+1-k)}B_k,\]

with $B_0=1$ and $B_1=-1/2$. Starting at $B_0$ is called the even index convention $(B_{2n+1}=0\ \rm{for}\ n>1)$.

Integer-overflow protection (IOP): for n > 35 the output is autoconverted to Rational{BigInt}. By default the capture message is activated: "Warning: bernoulliB converted to Rational{BigInt}".

Examples:

julia> o = [bernoulliB(n) for n=0:5]; println(o)
Rational{Int64}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1]

julia> bernoulliB(60)
Warning: bernoulliB converted to Rational{BigInt}
-1215233140483755572040304994079820246041491//56786730

julia> n = 60;
julia> bernoulliB(n) == bernoulliB_array(n)[end]             
true

CamMath.bernoulliB_array — Method

bernoulliB_array(nmax::T [; msg=true]) where {T<:Integer}

Bernoulli number array $[B_0,\cdots\ B_{nmax}]$, where nmax is the index of the highest Bernoulli number of the array (NB.: not the array length).

Examples:

julia> o = bernoulliB_array(8); println(o)
Rational{Int64}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1, 1//42, 0//1, -1//30]

julia> o = bernoulliB_array(big(8)); println(o)
Rational{BigInt}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1, 1//42, 0//1, -1//30]

julia> n = 60; msg = false;
julia>  bernoulliB(n; msg) == bernoulliB_array(n; msg)[end]            
true

Factorial

CamMath.bigfactorial — Method

bigfactorial(n::Int [; msg=true])

The product of all positive integers less than or equal to n,

\[n!=n(n-1)(n-2)⋯1.\]

In addition $0!=1$ by definition. For negative integers the factorial is zero. Integer-overflow protection (IOP): for n > 20 the output is autoconverted to BigInt.

Examples:

julia> bigfactorial(20) == factorial(20)
true

julia> bigfactorial(21)
Warning: bigfactorial converted to BigInt
51090942171709440000

julia> bigfactorial(21; msg=false)
51090942171709440000

julia> factorial(21)
ERROR: OverflowError: 21 is too large to look up in the table; consider using 
`factorial(big(21))` instead

Faulhaber polynomial

CamMath.faulhaber_polynom — Method

faulhaber_polynom(p::T; msg=true) where {T<:Integer}

Vector representation of the coefficients of the faulhaber_polynomial of degree p

\[ c=[c_0,⋯\ c_p]\]

where

\[ c_0=0,\ \ c_j=\frac{1}{p}{\binom{p}{p-j}}B_{p-j},\]

with $j∈\{ 1,⋯\ p\}$. The $B_0,⋯\ B_{p-1}$ are Bernoulli numbers (but with $B_1=+\frac{1}{2}$ rather than $-\frac{1}{2}$). Integer-overflow protection: for p > 36 the output is autoconverted to Rational{BigInt}. By default the capture message is activated: "Warning: faulhaber_polynom converted to Rational{BigInt}".

Example:

faulhaber_polynom(6)
7-element Vector{Rational{Int64}}:
  0//1
  0//1
 -1//12
  0//1
  5//12
  1//2
  1//6

CamMath.faulhaber_polynomial — Method

faulhaber_polynomial(n::T, p::Int; msg=true) where {T<:Integer}

Faulhaber polynomial of degree p

\[ F(n,p)=\sum_{j=0}^{p}c_{j}n^{j},\]

where the coefficients are contained in the coefficient vector c=[c_0,⋯\ c_p] given by faulhaber_polynom. Integer-overflow protection: in the event of integer overflow the output is converted to Rational{BigInt} and (by default) the capture message is activated: "Warning: faulhaber_polynomial converted to Rational{BigInt}".

Example:

julia> faulhaber_polynomial(3, 6)
276

julia> faulhaber_polynomial(5, 30)
Warning: faulhaber_polynomial converted to Rational{BigInt}
186552813930161650665

CamMath.faulhaber_summation — Method

faulhaber_summation(n::T, p::Int) where {T<:Integer}

Sum of the $p^{th}$ power of the first $n$ natural numbers

\[ \sum_{k=1}^{n}k^{p}=F(n,p+1).\]

where $F(n,p)$ is the faulhaber_polynomial of degree p.

Examples:

julia> faulhaber_summation(3,5)
276

julia> faulhaber_summation(3,60)
Warning: faulhaber_polynom converted to Rational{BigInt}
42391158276369125018901280178

HarmonicNumber

CamMath.harmonicNumber — Method

harmonicNumber(n::T [; msg=true]) where {T<:Integer}

Sum of the reciprocals of the first $n$ natural numbers

\[ H_n=\sum_{k=1}^{n}\frac{1}{k}.\]

Integer-overflow protection: for n > 46 the output is autoconverted to Rational{BigInt}. By default the capture message is activated: "Warning: harmonicNumber autoconverted to Rational{BigInt}".

Examples:

julia> o = harmonicNumber_array(9); println(o)
Rational{Int64}[1//1, 3//2, 11//6, 25//12, 137//60, 49//20, 363//140, 761//280, 7129//2520]

julia> o = [harmonicNumber(46; msg=true)]; println(o)
Rational{Int64}[5943339269060627227//1345655451257488800]

julia> o = [harmonicNumber(47; msg=true)]; println(o)
Warning: harmonicNumber autoconverted to Rational{BigInt}
Rational{BigInt}[282057509927739620069//63245806209101973600]

julia> harmonicNumber(12) == harmonicNumber(12, 1)
true

CamMath.harmonicNumber_array — Method

harmonicNumber_array(nmax::T [; msg=true]) where {T<:Integer}

Sum of the reciprocals of the first $n$ natural numbers

\[ H_n=\sum_{k=1}^{n}\frac{1}{k}.\]

Integer-overflow protection: for n > 46 the output is autoconverted to Rational{BigInt}. By default the capture message is activated: "Warning: harmonicNumber autoconverted to Rational{BigInt}".

Examples:

julia> o = harmonicNumber_array(9); println(o)
Rational{Int64}[1//1, 3//2, 11//6, 25//12, 137//60, 49//20, 363//140, 761//280, 7129//2520]

julia> o = [harmonicNumber(46; msg=true)]; println(o)
Rational{Int64}[5943339269060627227//1345655451257488800]

julia> o = [harmonicNumber(47; msg=true)]; println(o)
Warning: harmonicNumber autoconverted to Rational{BigInt}
Rational{BigInt}[282057509927739620069//63245806209101973600]

julia> harmonicNumber(12) == harmonicNumber(12, 1)
true

CamMath.harmonicNumber — Method

harmonicNumber(n::T, p::Int [; msg=true]) where {T<:Integer}

Sum of the $p^{th}$ power of reciprocals of the first $n$ numbers

\[ H_{n,p}=\sum_{k=1}^{n}\frac{1}{k^p}.\]

Integer-overflow protection: the output is autoconverted to Rational{BigInt} when required. By default the capture message is activated: "Warning: harmonicNumber autoconverted to Rational{BigInt}".

Examples:

julia> o = [harmonicNumber(46,1; msg=true)]; println(o)
Rational{Int64}[5943339269060627227//1345655451257488800]

julia> o = [harmonicNumber(47,1; msg=true)]; println(o)
Warning: harmonicNumber autoconverted to Rational{BigInt}"
Rational{BigInt}[280682601097106968469//63245806209101973600]

julia> o = [harmonicNumber(47,1)]; println(o)
Rational{BigInt}[280682601097106968469//63245806209101973600]

harmonicNumber(12, -3) == faulhaber_summation(12, 3)
  true

Pascal triangle

CamMath.pascal_triangle — Method

pascal_triangle(nmax [, T=Int])

Pascal triangle of binomial coefficients $\binom{n}{k}$ for $n=0,\ 1,⋯\ nmax$

Example:

pascal_triangle(5)
6-element Vector{Vector{Int64}}:
 [1]
 [1, 1]
 [1, 2, 1]
 [1, 3, 3, 1]
 [1, 4, 6, 4, 1]
 [1, 5, 10, 10, 5, 1]

CamMath.pascal_next — Method

pascal_next(nmax)

Next row of Pascal triangle

Example:

a = [1, 4, 6, 4, 1]
pascal_next(a)
 [1, 5, 10, 10, 5, 1]