CamiMath.jl

Mathematics library with integer-overload protection

Table of contents

CamiMath.jl

Table of contents

Bernoulli number

Factorial

Faulhaber polynomial

HarmonicNumber

Pascal triangle

Bernoulli number

CamiMath.bernoulliB — Method

bernoulliB(n::T [; msg=true]) where {T<:Integer}

Bernoulli numbers of index n are defined by the recurrence relation

\[ B_n = - \frac{1}{n+1}\sum_{k=0}^{n-1}\frac{(n+1)!}{k!(n+1-k)}B_k,\]

with $B_0=1$ and $B_1=-1/2$. Including $B_0$ results in the even index convention $(B_{2n+1}=0$ for $n>1)$.

Integer overflow protection (IOP): for n > 35 the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Examples:

julia> o = [bernoulliB(n) for n=0:5]; println(o)
Rational{Int64}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1]

julia> bernoulliB(60)
IOP capture: bernoulliB(60) converted to Rational{BigInt}
-1215233140483755572040304994079820246041491//56786730

julia> n = 60;
julia> bernoulliB(n) == bernoulliB_array(n)[end]             
true

CamiMath.bernoulliB_array — Method

bernoulliB_array(nmax::T [; msg=true]) where {T<:Integer}

Bernoulli number array $[B_0,\cdots\ B_{nmax}]$, where nmax is the index of the highest Bernoulli number of the array (NB.: not the array length).

Examples:

julia> o = bernoulliB_array(8); println(o)
Rational{Int64}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1, 1//42, 0//1, -1//30]

julia> o = bernoulliB_array(big(8)); println(o)
Rational{BigInt}[1//1, -1//2, 1//6, 0//1, -1//30, 0//1, 1//42, 0//1, -1//30]

julia> n = 60; msg = false;
julia>  bernoulliB(n; msg) == bernoulliB_array(n; msg)[end]            
true

Factorial

CamiMath.bigfactorial — Method

bigfactorial(n::Int [; msg=true])

The product of all positive integers less than or equal to n,

\[n!=n(n-1)(n-2)⋯1.\]

In addition $0!=1$ by definition. For negative integers the factorial is zero.

Integer overflow protection (IOP): for n > 20 the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Examples:

julia> bigfactorial(20) == factorial(20)
true

julia> bigfactorial(21)
IOP capture: bigfactorial(21) converted to BigInt
51090942171709440000

julia> bigfactorial(21; msg=false)
51090942171709440000

julia> factorial(21)
ERROR: OverflowError: 21 is too large to look up in the table; consider using 
`factorial(big(21))` instead

Faulhaber polynomial

CamiMath.faulhaber_polynom — Method

faulhaber_polynom(p::T [; msg=true]) where {T<:Integer}

Vector representation of the coefficients of the faulhaber_polynomial of degree p

\[ c=[c_0,⋯\ c_p]\]

where $c_0=0,\ \ c_j=\frac{1}{p}{\binom{p}{p-j}}B_{p-j}$, with $j∈\{ 1,⋯\ p\}$. The $B_{p-j}$ are BernoulliB in the even index convention (but with $B_1=+\frac{1}{2}$ rather than $-\frac{1}{2}$).

Integer overflow protection (IOP): for p > 36 the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Example:

faulhaber_polynom(6)
7-element Vector{Rational{Int64}}:
  0//1
  0//1
 -1//12
  0//1
  5//12
  1//2
  1//6

CamiMath.faulhaber_polynomial — Method

faulhaber_polynomial(n::T, p::Int; msg=true) where {T<:Integer}

Faulhaber polynomial of degree p

\[ F(n,p)=\sum_{j=0}^{p}c_{j}n^{j},\]

where n is a positive integer and the coefficients are contained in the vector $c=[c_0,⋯\ c_p]$ given by faulhaber_polynom.

Integer overflow protection (IOP): on integer overflow the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Example:

julia> faulhaber_polynomial(3, 6)
276

julia> faulhaber_polynomial(5, 30)
IOP capture: faulhaber_polynomial(5, 30) autoconverted to Rational{BigInt}
186552813930161650665

CamiMath.faulhaber_summation — Method

faulhaber_summation(n::T, p::Int) where {T<:Integer}

Sum of the $p^{th}$ power of the first $n$ natural numbers

\[ \sum_{k=1}^{n}k^{p}=H_{n,-p}=F(n,p+1).\]

where $H_{n,-p}$ is a harmonicNumber of power -p and $F(n,p)$ a faulhaber_polynomial of power p.

Integer overflow protection (IOP): on integer overflow the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Examples:

julia> faulhaber_summation(3,5)
276

julia> faulhaber_summation(3,60)
IOP capture: faulhaber_polynom autoconverted to Rational{BigInt}
42391158276369125018901280178

HarmonicNumber

CamiMath.harmonicNumber — Method

harmonicNumber(n::T [; msg=true]) where {T<:Integer}

Sum of the reciprocals of the first $n$ natural numbers

\[ H_n=\sum_{k=1}^{n}\frac{1}{k}.\]

Integer overflow protection (IOP): for n > 46 the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Examples:

julia> o = harmonicNumber_array(8); println(o)
Rational{Int64}[1//1, 3//2, 11//6, 25//12, 137//60, 49//20, 363//140, 761//280]

julia> harmonicNumber(46)
5943339269060627227//1345655451257488800

julia> harmonicNumber(47)
IOP capture: harmonicNumber(47) converted to Rational{BigInt}
280682601097106968469//63245806209101973600

julia> harmonicNumber(12) == harmonicNumber(12, 1)
true

CamiMath.harmonicNumber_array — Method

harmonicNumber_array(nmax::T [; msg=true]) where {T<:Integer}

Sum of the reciprocals of the first $n$ natural numbers

\[ H_n=\sum_{k=1}^{n}\frac{1}{k}.\]

Integer overflow protection (IOP): for nmax > 46 the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Examples:

julia> o = harmonicNumber_array(8); println(o)
Rational{Int64}[1//1, 3//2, 11//6, 25//12, 137//60, 49//20, 363//140, 761//280] 

julia> harmonicNumber(12) == harmonicNumber(12, 1)
true

CamiMath.harmonicNumber — Method

harmonicNumber(n::T, p::Int [; msg=true]) where {T<:Integer}

Sum of the $p^{th}$ power of reciprocals of the first $n$ positive integers,

\[ H_{n,p}=\sum_{k=1}^{n}\frac{1}{k^p}.\]

Integer overflow protection (IOP): on integer overflow the output is converted to Rational{BigInt}. By default the IOP capture message is activated.

Examples:

julia> harmonicNumber(46, 1)
5943339269060627227//1345655451257488800

julia> harmonicNumber(47, 1)
IOP capture: harmonicNumber(47, 1) converted to Rational{BigInt}
280682601097106968469//63245806209101973600

harmonicNumber(12, -3) == faulhaber_summation(12, 3)
  true

Pascal triangle

CamiMath.pascal_triangle — Method

pascal_triangle(nmax [, T=Int])

Pascal triangle of binomial coefficients $\binom{n}{k}$ for $n=0,\ 1,⋯\ nmax$

Example:

pascal_triangle(5)
6-element Vector{Vector{Int64}}:
 [1]
 [1, 1]
 [1, 2, 1]
 [1, 3, 3, 1]
 [1, 4, 6, 4, 1]
 [1, 5, 10, 10, 5, 1]

CamiMath.pascal_next — Method

pascal_next(nmax)

Next row of Pascal triangle

Example:

a = [1, 4, 6, 4, 1]
pascal_next(a)
 [1, 5, 10, 10, 5, 1]